Change mode-ton dtype to int64 to prevent overflow

+++ Scripts/utils/freight_utils.py
@@ -63,10 +63,10 @@ def write_domestic_leg_summary(demand_trade: dict, impedance: dict,
     for purpose in demand_trade:
         trade_type = "export" if purpose.is_export else "import"
         modes = list(demand_trade[purpose])
-        mode_tons = [numpy.sum(demand_trade[purpose][mode], dtype=int)
+        mode_tons = [numpy.sum(demand_trade[purpose][mode], dtype=numpy.int32)
                      for mode in modes]
         mode_ton_dist = [numpy.sum(demand_trade[purpose][mode]
-                                   * impedance[mode]["dist"], dtype=int)
+                                   * impedance[mode]["dist"], dtype=numpy.int64)
                         for mode in modes]
         df_data.append(DataFrame(data={
             "Commodity": [purpose.name] * len(modes),
